requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
flask-migrate==3.1.0
python-dotenv==0.19.1
sqlalchemy==1.4.23
werkzeug==2.0.1
